copias: gosec limpio, con la guarda de rutas anotada donde de verdad muerde
gosec es BLOQUEANTE en ci.yml y sacaba cinco hallazgos del paquete nuevo, todos
con confianza alta. Cuatro son G703 (path traversal por analisis de taint) y uno
es G115 (desbordamiento al convertir uint64 a int).

El G115 se ARREGLA, no se anota: `int(e.Borrado.Entrada) >= len(...)` da la
vuelta en una maquina de 32 bits, y un escenario que borrara la entrada 2^32
pasaria por "borra la entrada 0". La comparacion se hace en uint64.

De los cuatro G703, tres son escrituras a rutas que fija el propio ensayo o que
teclea el operador en su maquina, con nombres constantes. Van con #nosec y su
motivo, como el resto del repositorio.

El cuarto es el de verdad, y su anotacion dice por que no es un hallazgo VIVO en
vez de por que se ignora: es la escritura de Restaurar, la unica del paquete
cuyo nombre viene de entrada de un tercero (el manifiesto de la replica), y en
ese mismo bucle pasa antes por nombreDeArtefactoValido. Lo que el analisis de
taint no ve lo ve un test, y esta escrito a que test mirar si alguien quita la
guarda.

Comprobado ademas: govulncheck sin hallazgos.

+++ herramientas/ensayocopia/sembrar.go
@@ -67,7 +67,10 @@ func cargarEscenario() (Escenario, error) {
 		return e, fmt.Errorf("el escenario no declara base legal del borrado, y un borrado sin " +
 			"base legal no es un borrado legal: el nucleo lo rechaza al firmarlo")
 	}
-	if int(e.Borrado.Entrada) >= len(e.Entradas) {
+	// La comparacion se hace en uint64 y no metiendo el indice en un int: en
+	// una maquina de 32 bits, int(unNumeroGrande) da la vuelta y un escenario
+	// que borra la entrada 2^32 pasaria por "borra la entrada 0".
+	if e.Borrado.Entrada >= uint64(len(e.Entradas)) {
 		return e, fmt.Errorf("el escenario borra la entrada %d y solo hay %d",
 			e.Borrado.Entrada, len(e.Entradas))
 	}
